update tests and docs
