--- titiler/api/api_v1/__init__.py
@@ -1,1 +1,0 @@
-"""titiler.api.api_v1"""
