+++ titiler/api/endpoints/stac.py
@@ -1,6 +1,4 @@
-"""API metadata."""
-
-from typing import Any, Dict, Optional
+"""API for SpatioTemporal Asset Catalog items."""
 
 import os
 from io import BytesIO
@@ -8,15 +6,20 @@ from urllib.parse import urlencode
 
 import numpy
 
-from rio_tiler_crs import COGReader
+from typing import Dict, Union, List
 
-from fastapi import APIRouter, Depends, Query, Path
+from stac_tiler import STACReader
+
+from fastapi import APIRouter, Depends, Query, Path, HTTPException
+from starlette import status
 from starlette.requests import Request
 from starlette.responses import Response, HTMLResponse
-from starlette.templating import Jinja2Templates
 
 from titiler.models.metadata import cogBounds, cogInfo, cogMetadata
 from titiler.api.deps import CommonMetadataParams
+
+from typing import Any, Optional
+
 
 from rasterio.transform import from_bounds
 from rio_tiler.utils import render
@@ -25,58 +28,64 @@ from rio_tiler.profiles import img_profiles
 from titiler.api import utils
 from titiler.api.deps import CommonImageParams, TileMatrixSetNames, morecantile
 from titiler.db.memcache import CacheLayer
-from titiler.ressources.enums import ImageType, ImageMimeTypes, MimeTypes
+from titiler.ressources.enums import ImageType, ImageMimeTypes
 from titiler.ressources.common import drivers
-from titiler.ressources.responses import TileResponse, XMLResponse
+from titiler.ressources.responses import TileResponse
 from titiler.models.mapbox import TileJSON
 from titiler.templates.factory import web_template
 
 router = APIRouter()
-templates = Jinja2Templates(directory="titiler/templates")
 
 
 @router.get(
     "/bounds",
     response_model=cogBounds,
-    responses={200: {"description": "Return the bounds of the COG."}},
+    responses={200: {"description": "Return the bounds of the STAC item."}},
 )
-async def cog_bounds(
-    resp: Response, url: str = Query(..., description="Cloud Optimized GeoTIFF URL."),
+async def stac_bounds(
+    resp: Response, url: str = Query(..., description="STAC item URL."),
 ):
-    """Return the bounds of the COG."""
+    """Return the bounds of the STAC item."""
     resp.headers["Cache-Control"] = "max-age=3600"
-    with COGReader(url) as cog:
-        return {"bounds": cog.bounds}
+    with STACReader(url) as stac:
+        return {"bounds": stac.bounds}
 
 
 @router.get(
     "/info",
-    response_model=cogInfo,
-    response_model_exclude={"minzoom", "maxzoom", "center"},
+    response_model=Union[List[str], Dict[str, cogInfo]],
+    response_model_exclude={"__all__": {"minzoom", "maxzoom", "center"}},
     response_model_exclude_none=True,
-    responses={200: {"description": "Return basic info on COG."}},
+    responses={200: {"description": "Return basic info for STAC item's assets"}},
 )
-async def cog_info(
-    resp: Response, url: str = Query(..., description="Cloud Optimized GeoTIFF URL.")
+async def stac_info(
+    resp: Response,
+    url: str = Query(..., description="STAC item URL."),
+    assets: str = Query(None, description="comma (,) separated list of asset names."),
 ):
-    """Return basic info on COG."""
+    """Return basic info on STAC item's COG."""
     resp.headers["Cache-Control"] = "max-age=3600"
-    with COGReader(url) as cog:
-        info = cog.info
+    with STACReader(url) as stac:
+        if not assets:
+            return stac.assets
+
+        info = stac.info(assets.split(","))
+
     return info
 
 
 @router.get(
     "/metadata",
-    response_model=cogMetadata,
-    response_model_exclude={"minzoom", "maxzoom", "center"},
+    response_model=Dict[str, cogMetadata],
+    response_model_exclude={"__all__": {"minzoom", "maxzoom", "center"}},
     response_model_exclude_none=True,
-    responses={200: {"description": "Return the metadata of the COG."}},
+    responses={200: {"description": "Return the metadata for STAC item's assets."}},
 )
-async def cog_metadata(
+async def stac_metadata(
     request: Request,
     resp: Response,
-    url: str = Query(..., description="Cloud Optimized GeoTIFF URL."),
+    url: str = Query(..., description="STAC item URL."),
+    assets: str = Query(..., description="comma (,) separated list of asset names."),
     metadata_params: CommonMetadataParams = Depends(CommonMetadataParams),
 ):
     """Return the metadata of the COG."""
@@ -89,10 +98,11 @@ async def cog_metadata(
     kwargs.pop("max_size", None)
     kwargs.pop("histogram_bins", None)
     kwargs.pop("histogram_range", None)
+    kwargs.pop("assets", None)
 
-    with COGReader(url) as cog:
-        info = cog.info
-        stats = cog.stats(
+    with STACReader(url) as stac:
+        info = stac.metadata(
+            assets.split(","),
             metadata_params.pmin,
             metadata_params.pmax,
             nodata=metadata_params.nodata,
@@ -101,7 +111,6 @@ async def cog_metadata(
             hist_options=metadata_params.hist_options,
             **kwargs,
         )
-        info["statistics"] = stats
 
     resp.headers["Cache-Control"] = "max-age=3600"
     return info
@@ -132,7 +141,7 @@ params: Dict[str, Any] = {
 @router.get(r"/tiles/{TileMatrixSetId}/{z}/{x}/{y}\.{ext}", **params)
 @router.get(r"/tiles/{TileMatrixSetId}/{z}/{x}/{y}@{scale}x", **params)
 @router.get(r"/tiles/{TileMatrixSetId}/{z}/{x}/{y}@{scale}x\.{ext}", **params)
-async def cog_tile(
+async def stac_tile(
     z: int = Path(..., ge=0, le=30, description="Mercator tiles's zoom level"),
     x: int = Path(..., description="Mercator tiles's column"),
     y: int = Path(..., description="Mercator tiles's row"),
@@ -144,13 +153,20 @@ async def cog_tile(
         1, gt=0, lt=4, description="Tile size scale. 1=256x256, 2=512x512..."
     ),
     ext: ImageType = Query(None, description="Output image type. Default is auto."),
-    url: str = Query(..., description="Cloud Optimized GeoTIFF URL."),
+    url: str = Query(..., description="STAC Item URL."),
+    assets: str = Query("", description="comma (,) separated list of asset names."),
     image_params: CommonImageParams = Depends(CommonImageParams),
     cache_client: CacheLayer = Depends(utils.get_cache),
 ):
-    """Create map tile from a COG."""
+    """Create map tile from a STAC item."""
     timings = []
     headers: Dict[str, str] = {}
+
+    if not image_params.expression and not assets:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Must pass Expression or Asset list.",
+        )
 
     tile_hash = utils.get_hash(
         **dict(
@@ -181,18 +197,17 @@ async def cog_tile(
 
     if not content:
         with utils.Timer() as t:
-            with COGReader(url, tms=tms) as cog:
-                tile, mask = cog.tile(
+            with STACReader(url, tms=tms) as stac:
+                tile, mask = stac.tile(
                     x,
                     y,
                     z,
+                    assets=assets.split(","),
                     tilesize=tilesize,
                     indexes=image_params.indexes,
                     expression=image_params.expression,
                     nodata=image_params.nodata,
                 )
-                colormap = image_params.color_map or cog.colormap
-
         timings.append(("Read", t.elapsed))
 
         if not ext:
@@ -221,7 +236,11 @@ async def cog_tile(
                     dst_transform = from_bounds(*bounds, tilesize, tilesize)
                     options = {"crs": tms.crs, "transform": dst_transform}
                 content = render(
-                    tile, mask, img_format=driver, colormap=colormap, **options
+                    tile,
+                    mask,
+                    img_format=driver,
+                    colormap=image_params.color_map,
+                    **options,
                 )
         timings.append(("Format", t.elapsed))
 
@@ -250,14 +269,20 @@ async def cog_tile(
     responses={200: {"description": "Return a tilejson"}},
     response_model_exclude_none=True,
 )
-async def cog_tilejson(
+async def stac_tilejson(
     request: Request,
     response: Response,
     TileMatrixSetId: TileMatrixSetNames = Query(
         TileMatrixSetNames.WebMercatorQuad,  # type: ignore
         description="TileMatrixSet Name (default: 'WebMercatorQuad')",
     ),
-    url: str = Query(..., description="Cloud Optimized GeoTIFF URL."),
+    url: str = Query(..., description="STAC Item URL."),
+    assets: str = Query("", description="comma (,) separated list of asset names."),
+    expression: Optional[str] = Query(
+        None,
+        title="Band Math expression",
+        description="rio-tiler's band math expression (e.g B1/B2)",
+    ),
     tile_format: Optional[ImageType] = Query(
         None, description="Output image type. Default is auto."
     ),
@@ -267,7 +292,7 @@ async def cog_tilejson(
     minzoom: Optional[int] = Query(None, description="Overwrite default minzoom."),
     maxzoom: Optional[int] = Query(None, description="Overwrite default maxzoom."),
 ):
-    """Return TileJSON document for a COG."""
+    """Return a TileJSON document for a STAC item."""
     scheme = request.url.scheme
     host = request.headers["host"]
 
@@ -276,22 +301,28 @@ async def cog_tilejson(
     kwargs.pop("tile_scale", None)
     kwargs.pop("TileMatrixSetId", None)
 
+    if not expression and not assets:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Expression or Assets HAVE to be set in the queryString.",
+        )
+
     qs = urlencode(list(kwargs.items()))
     if tile_format:
-        tile_url = f"{scheme}://{host}/cog/tiles/{TileMatrixSetId.name}/{{z}}/{{x}}/{{y}}@{tile_scale}x.{tile_format}?{qs}"
+        tile_url = f"{scheme}://{host}/stac/tiles/{TileMatrixSetId.name}/{{z}}/{{x}}/{{y}}@{tile_scale}x.{tile_format}?{qs}"
     else:
-        tile_url = f"{scheme}://{host}/cog/tiles/{TileMatrixSetId.name}/{{z}}/{{x}}/{{y}}@{tile_scale}x?{qs}"
+        tile_url = f"{scheme}://{host}/stac/tiles/{TileMatrixSetId.name}/{{z}}/{{x}}/{{y}}@{tile_scale}x?{qs}"
 
     tms = morecantile.tms.get(TileMatrixSetId.name)
-    with COGReader(url, tms=tms) as cog:
-        center = list(cog.center)
+    with STACReader(url, tms=tms) as stac:
+        center = list(stac.center)
         if minzoom:
             center[-1] = minzoom
         tjson = {
-            "bounds": cog.bounds,
+            "bounds": stac.bounds,
             "center": tuple(center),
-            "minzoom": minzoom or cog.minzoom,
-            "maxzoom": maxzoom or cog.maxzoom,
+            "minzoom": minzoom or stac.minzoom,
+            "maxzoom": maxzoom or stac.maxzoom,
             "name": os.path.basename(url),
             "tiles": [tile_url],
         }
@@ -300,75 +331,7 @@ async def cog_tilejson(
     return tjson
 
 
-@router.get("/WMTSCapabilities.xml", response_class=XMLResponse, tags=["OGC"])
-@router.get(
-    "/{TileMatrixSetId}/WMTSCapabilities.xml", response_class=XMLResponse, tags=["OGC"],
-)
-def wtms(
-    request: Request,
-    response: Response,
-    TileMatrixSetId: TileMatrixSetNames = Query(
-        TileMatrixSetNames.WebMercatorQuad,  # type: ignore
-        description="TileMatrixSet Name (default: 'WebMercatorQuad')",
-    ),
-    url: str = Query(..., description="Cloud Optimized GeoTIFF URL."),
-    tile_format: ImageType = Query(
-        ImageType.png, description="Output image type. Default is png."
-    ),
-    tile_scale: int = Query(
-        1, gt=0, lt=4, description="Tile size scale. 1=256x256, 2=512x512..."
-    ),
-):
-    """OGC WMTS endpoit."""
-    scheme = request.url.scheme
-    host = request.headers["host"]
-    endpoint = f"{scheme}://{host}/cog"
-
-    kwargs = dict(request.query_params)
-    kwargs.pop("tile_format", None)
-    kwargs.pop("tile_scale", None)
-    qs = urlencode(list(kwargs.items()))
-
-    tms = morecantile.tms.get(TileMatrixSetId.name)
-    with COGReader(url, tms=tms) as cog:
-        minzoom, maxzoom, bounds = cog.minzoom, cog.maxzoom, cog.bounds
-
-    media_type = ImageMimeTypes[tile_format.value].value
-
-    tileMatrix = []
-    for zoom in range(minzoom, maxzoom + 1):
-        matrix = tms.matrix(zoom)
-        tm = f"""
-                <TileMatrix>
-                    <ows:Identifier>{matrix.identifier}</ows:Identifier>
-                    <ScaleDenominator>{matrix.scaleDenominator}</ScaleDenominator>
-                    <TopLeftCorner>{matrix.topLeftCorner[0]} {matrix.topLeftCorner[1]}</TopLeftCorner>
-                    <TileWidth>{matrix.tileWidth}</TileWidth>
-                    <TileHeight>{matrix.tileHeight}</TileHeight>
-                    <MatrixWidth>{matrix.matrixWidth}</MatrixWidth>
-                    <MatrixHeight>{matrix.matrixHeight}</MatrixHeight>
-                </TileMatrix>"""
-        tileMatrix.append(tm)
-
-    tile_ext = f"@{tile_scale}x.{tile_format.value}"
-    return templates.TemplateResponse(
-        "wmts.xml",
-        {
-            "request": request,
-            "endpoint": endpoint,
-            "bounds": bounds,
-            "tileMatrix": tileMatrix,
-            "tms": tms,
-            "title": "Cloud Optimized GeoTIFF",
-            "query_string": qs,
-            "tile_format": tile_ext,
-            "media_type": media_type,
-        },
-        media_type=MimeTypes.xml.value,
-    )
-
-
 @router.get("/viewer", response_class=HTMLResponse, tags=["Webpage"])
-def cog_viewer(request: Request, template=Depends(web_template)):
-    """Cloud Optimized GeoTIFF Viewer."""
-    return template(request, "cog_index.html", "cog_tilejson", "cog_metadata")
+def stac_viewer(request: Request, template=Depends(web_template)):
+    """SpatioTemporal Asset Catalog Viewer."""
+    return template(request, "stac_index.html", "stac_tilejson", "stac_info")
